Organize image_cache module.
