+++ crates/zng-view/src/image_cache.rs
@@ -1,28 +1,28 @@
-use std::{fmt, sync::Arc};
-
-use image::GenericImageView as _;
 #[cfg(feature = "image_any")]
 use image::ImageDecoder as _;
+use std::{fmt, sync::Arc};
 
 use webrender::api::{ImageDescriptor, ImageDescriptorFlags, ImageFormat};
-use winit::{
-    event_loop::ActiveEventLoop,
-    window::{CustomCursor, Icon},
-};
 use zng_txt::ToTxt as _;
 use zng_txt::{Txt, formatx};
-#[cfg(feature = "image_any")]
-use zng_unit::PxDensityUnits as _;
 
-use zng_task::channel::{IpcBytes, IpcBytesMut, IpcReceiver};
-use zng_unit::{Px, PxDensity2d, PxPoint, PxSize};
+use zng_task::channel::{IpcBytes, IpcReceiver};
+use zng_unit::{Px, PxDensity2d, PxSize};
 use zng_view_api::{
     Event,
-    image::{ImageDataFormat, ImageId, ImageLoadedData, ImageMaskMode, ImageRequest},
+    image::{ImageDataFormat, ImageId, ImageLoadedData, ImageRequest},
 };
 
 use crate::{AppEvent, AppEventSender};
 use rustc_hash::FxHashMap;
+
+// Image data is provided to webrender directly from the BGRA8 shared memory.
+// The `ExternalImageId` is the Arc pointer to ImageData.
+mod capture;
+mod decode;
+mod encode;
+mod external;
+pub(crate) use external::{ImageUseMap, WrImageCache};
 
 #[cfg(not(feature = "image_any"))]
 mod lcms2 {
@@ -422,712 +422,6 @@ impl ImageCache {
         let _ = self.app_sender.send(AppEvent::Notify(Event::ImageLoaded(data)));
     }
 
-    #[cfg(feature = "image_any")]
-    fn header_decode(fmt: &ImageDataFormat, data: &[u8]) -> Result<ImageHeader, Txt> {
-        let maybe_fmt = match fmt {
-            ImageDataFormat::FileExtension(ext) => image::ImageFormat::from_extension(ext.as_str()),
-            ImageDataFormat::MimeType(t) => t.strip_prefix("image/").and_then(image::ImageFormat::from_extension),
-            ImageDataFormat::Unknown => None,
-            ImageDataFormat::Bgra8 { .. } => unreachable!(),
-            ImageDataFormat::A8 { .. } => unreachable!(),
-            _ => None,
-        };
-
-        let reader = match maybe_fmt {
-            Some(fmt) => image::ImageReader::with_format(std::io::Cursor::new(data), fmt),
-            None => image::ImageReader::new(std::io::Cursor::new(data))
-                .with_guessed_format()
-                .map_err(|e| e.to_txt())?,
-        };
-
-        match reader.format() {
-            Some(fmt) => {
-                use image::metadata::Orientation::*;
-
-                let mut decoder = match reader.into_decoder() {
-                    Ok(d) => d,
-                    Err(e) => {
-                        // decoder error, try fallback to Unknown
-                        if let image::ImageError::Decoding(_) = &e
-                            && maybe_fmt.is_some()
-                            && let Ok(r) = Self::header_decode(&ImageDataFormat::Unknown, data)
-                        {
-                            return Ok(r);
-                        }
-                        return Err(e.to_txt());
-                    }
-                };
-                let (mut w, mut h) = decoder.dimensions();
-                let orientation = decoder.orientation().unwrap_or(NoTransforms);
-                if matches!(orientation, Rotate90 | Rotate270 | Rotate90FlipH | Rotate270FlipH) {
-                    std::mem::swap(&mut w, &mut h)
-                }
-
-                let mut density = None;
-                #[cfg(feature = "image_png")]
-                let mut png_gamma = None;
-                #[cfg(feature = "image_png")]
-                let mut png_chromaticities = None;
-
-                match fmt {
-                    #[cfg(feature = "image_jpeg")]
-                    image::ImageFormat::Jpeg => {
-                        // `image` uses `zune-jpeg`, that decoder does not parse density correctly,
-                        // so we do it manually here
-                        fn parse_density(data: &[u8]) -> Option<(u8, u16, u16)> {
-                            let mut i = 0;
-                            while i + 4 < data.len() {
-                                // APP0
-                                if data[i] == 0xFF && data[i + 1] == 0xE0 {
-                                    let len = u16::from_be_bytes([data[i + 2], data[i + 3]]) as usize;
-                                    if i + 2 + len > data.len() {
-                                        break;
-                                    }
-
-                                    // APP0 payload starts at i+4, identifier is 5 bytes: "JFIF\0"
-                                    let p = i + 4;
-                                    if &data[p..p + 5] == b"JFIF\0" && p + 14 <= data.len() {
-                                        let unit = data[p + 7];
-                                        let x = u16::from_be_bytes([data[p + 8], data[p + 9]]);
-                                        let y = u16::from_be_bytes([data[p + 10], data[p + 11]]);
-                                        return Some((unit, x, y));
-                                    }
-
-                                    i += 2 + len;
-                                } else if data[i] == 0xFF && data[i + 1] == 0xDA {
-                                    // Start of Scan
-                                    break;
-                                } else {
-                                    i += 1;
-                                }
-                            }
-                            None
-                        }
-                        if let Some((unit, x, y)) = parse_density(data) {
-                            match unit {
-                                // inches
-                                1 => {
-                                    density = Some(PxDensity2d::new((x as f32).ppi(), (y as f32).ppi()));
-                                }
-                                // centimeters
-                                2 => {
-                                    density = Some(PxDensity2d::new((x as f32).ppcm(), (y as f32).ppcm()));
-                                }
-                                _ => {}
-                            }
-                        }
-                    }
-                    #[cfg(feature = "image_png")]
-                    image::ImageFormat::Png => {
-                        let d = png::Decoder::new_with_limits(std::io::Cursor::new(data), png::Limits { bytes: usize::MAX });
-                        let d = d.read_info().map_err(|e| e.to_txt())?;
-                        let info = d.info();
-                        if let Some(d) = info.pixel_dims {
-                            match d.unit {
-                                png::Unit::Unspecified => {}
-                                png::Unit::Meter => {
-                                    use zng_unit::PxDensity;
-
-                                    density = Some(PxDensity2d::new(
-                                        PxDensity::new_ppm(d.xppu as f32),
-                                        PxDensity::new_ppm(d.yppu as f32),
-                                    ));
-                                }
-                            }
-                        }
-                        png_gamma = info.gama_chunk;
-                        png_chromaticities = info.chrm_chunk;
-                    }
-                    #[cfg(feature = "image_tiff")]
-                    image::ImageFormat::Tiff => {
-                        use tiff::{decoder::ifd::Value, tags::Tag};
-                        let mut d = tiff::decoder::Decoder::new(std::io::Cursor::new(data))
-                            .map_err(|e| e.to_txt())?
-                            .with_limits(tiff::decoder::Limits::unlimited());
-                        let res_unit = d.get_tag(Tag::ResolutionUnit).ok().and_then(|t| t.into_u16().ok()).unwrap_or(2);
-                        if let Ok(Value::Rational(x_num, x_denom)) = d.get_tag(Tag::XResolution)
-                            && let Ok(Value::Rational(y_num, y_denom)) = d.get_tag(Tag::YResolution)
-                        {
-                            let x = x_num as f32 / x_denom as f32;
-                            let y = y_num as f32 / y_denom as f32;
-                            match res_unit {
-                                // inches
-                                2 => {
-                                    density = Some(PxDensity2d::new(x.ppi(), y.ppi()));
-                                }
-                                // centimeters
-                                3 => {
-                                    density = Some(PxDensity2d::new(x.ppcm(), y.ppcm()));
-                                }
-                                _ => {}
-                            }
-                        }
-                    }
-                    _ => {}
-                }
-
-                if density.is_none()
-                    && let Ok(Some(exif)) = decoder.exif_metadata()
-                    && let Ok(exif) = exif::Reader::new().read_raw(exif)
-                {
-                    use exif::Tag;
-                    if let Some(unit) = exif.get_field(Tag::ResolutionUnit, exif::In::PRIMARY)
-                        && let Some(x) = exif.get_field(Tag::XResolution, exif::In::PRIMARY)
-                        && let Some(y) = exif.get_field(Tag::YResolution, exif::In::PRIMARY)
-                        && let exif::Value::Rational(x) = &x.value
-                        && let exif::Value::Rational(y) = &y.value
-                    {
-                        let x = x[0].to_f32();
-                        let y = y[0].to_f32();
-                        match unit.value.get_uint(0) {
-                            // inches
-                            Some(2) => density = Some(PxDensity2d::new(x.ppi(), y.ppi())),
-                            // centimeters
-                            Some(3) => density = Some(PxDensity2d::new(x.ppcm(), y.ppcm())),
-                            _ => {}
-                        }
-                    }
-                }
-
-                let mut icc_profile = None;
-                if let Ok(Some(icc)) = decoder.icc_profile() {
-                    match lcms2::Profile::new_icc(&icc) {
-                        Ok(p) => icc_profile = Some(p),
-                        Err(e) => tracing::error!("error parsing ICC profile, {e}"),
-                    }
-                }
-                #[cfg(feature = "image_png")]
-                if icc_profile.is_none() {
-                    // PNG has some color management metadata, convert to standard
-                    icc_profile = crate::util::png_color_metadata_to_icc(png_gamma, png_chromaticities);
-                }
-
-                Ok(ImageHeader {
-                    format: fmt,
-                    size: PxSize::new(Px(w as i32), Px(h as i32)),
-                    orientation,
-                    density,
-                    icc_profile,
-                })
-            }
-            None => Err(Txt::from_static("unknown format")),
-        }
-    }
-
-    #[cfg(feature = "image_any")]
-    fn image_decode(
-        buf: &[u8],
-        format: image::ImageFormat,
-        downscale: Option<zng_view_api::image::ImageDownscale>,
-        orientation: image::metadata::Orientation,
-    ) -> image::ImageResult<image::DynamicImage> {
-        let buf = std::io::Cursor::new(buf);
-
-        let mut reader = image::ImageReader::new(buf);
-        reader.set_format(format);
-        reader.no_limits();
-        let mut image = reader.decode()?;
-
-        image.apply_orientation(orientation);
-
-        if let Some(s) = downscale {
-            let (img_w, img_h) = (image.width(), image.height());
-            use zng_view_api::image::ImageDownscale::*;
-            match s {
-                Fit(s) => {
-                    let w = img_w.min(s.width.0 as u32);
-                    let h = img_h.min(s.height.0 as u32);
-                    if w != img_w || h != img_h {
-                        image = image.resize(w, h, image::imageops::FilterType::Triangle);
-                    }
-                }
-                Fill(s) => {
-                    let w = img_w.min(s.width.0 as u32);
-                    let h = img_h.min(s.height.0 as u32);
-                    if w != img_w && h != img_h {
-                        image = image.resize_to_fill(w, h, image::imageops::FilterType::Triangle);
-                    }
-                }
-            }
-        }
-
-        Ok(image)
-    }
-
-    fn convert_decoded(
-        image: image::DynamicImage,
-        mask: Option<ImageMaskMode>,
-        density: Option<PxDensity2d>,
-        icc_profile: Option<lcms2::Profile>,
-    ) -> std::io::Result<RawLoadedImg> {
-        use image::DynamicImage::*;
-
-        let mut is_opaque = true;
-        let size = image.dimensions();
-        let pixels_len = size.0 as usize * size.1 as usize;
-
-        let mut pixels = match image {
-            ImageLuma8(img) => {
-                let raw = img.into_raw();
-                if mask.is_some() {
-                    is_opaque = !raw.iter().any(|&a| a < 255);
-                    IpcBytesMut::from_vec_blocking(raw)?
-                } else {
-                    let mut bgra = IpcBytes::new_mut_blocking(pixels_len * 4)?;
-                    for (p, l) in bgra.chunks_exact_mut(4).zip(raw) {
-                        p.copy_from_slice(&[l, l, l, 255])
-                    }
-                    bgra
-                }
-            }
-            ImageLumaA8(img) => {
-                let raw = img.into_raw();
-                if let Some(mask) = mask {
-                    let mut a = IpcBytes::new_mut_blocking(pixels_len)?;
-                    match mask {
-                        ImageMaskMode::A => {
-                            for (p, la) in a.iter_mut().zip(raw.chunks_exact(2)) {
-                                if la[1] < 255 {
-                                    is_opaque = false;
-                                }
-                                *p = la[1];
-                            }
-                        }
-                        ImageMaskMode::B | ImageMaskMode::G | ImageMaskMode::R | ImageMaskMode::Luminance => {
-                            for (p, la) in a.iter_mut().zip(raw.chunks_exact(2)) {
-                                if la[0] < 255 {
-                                    is_opaque = false;
-                                }
-                                *p = la[0];
-                            }
-                        }
-                        _ => unimplemented!(),
-                    }
-                    a
-                } else {
-                    let mut bgra = IpcBytes::new_mut_blocking(pixels_len * 4)?;
-                    for (p, la) in bgra.chunks_exact_mut(4).zip(raw.chunks_exact(2)) {
-                        p.copy_from_slice(&if la[1] < 255 {
-                            is_opaque = false;
-                            let l = la[0] as f32 * la[1] as f32 / 255.0;
-                            let l = l as u8;
-                            [l, l, l, la[1]]
-                        } else {
-                            let l = la[0];
-                            [l, l, l, la[1]]
-                        });
-                    }
-                    bgra
-                }
-            }
-            ImageRgb8(img) => {
-                let raw = img.into_raw();
-                if let Some(mask) = mask {
-                    let mut a = IpcBytes::new_mut_blocking(pixels_len)?;
-                    match mask {
-                        ImageMaskMode::Luminance | ImageMaskMode::A => {
-                            for (p, rgb) in a.iter_mut().zip(raw.chunks(3)) {
-                                let l = luminance(rgb);
-                                if l < 255 {
-                                    is_opaque = false;
-                                }
-                                *p = l;
-                            }
-                        }
-                        mask => {
-                            let channel = match mask {
-                                ImageMaskMode::B => 2,
-                                ImageMaskMode::G => 1,
-                                ImageMaskMode::R => 0,
-                                _ => unreachable!(),
-                            };
-                            for (p, rgb) in a.iter_mut().zip(raw.chunks(3)) {
-                                let c = rgb[channel];
-                                if c < 255 {
-                                    is_opaque = false;
-                                }
-                                *p = c;
-                            }
-                        }
-                    }
-                    a
-                } else {
-                    let mut bgra = IpcBytes::new_mut_blocking(pixels_len * 4)?;
-                    for (p, rgb) in bgra.chunks_exact_mut(4).zip(raw.chunks_exact(3)) {
-                        p.copy_from_slice(&[rgb[2], rgb[1], rgb[0], 255]);
-                    }
-                    bgra
-                }
-            }
-            ImageRgba8(img) => {
-                let raw = img.into_raw();
-                if let Some(mask) = mask {
-                    let mut a = IpcBytes::new_mut_blocking(pixels_len)?;
-                    match mask {
-                        ImageMaskMode::Luminance => {
-                            for (p, rgba) in a.iter_mut().zip(raw.chunks_exact(4)) {
-                                let c = luminance(&rgba[..3]);
-                                if c < 255 {
-                                    is_opaque = false;
-                                }
-                                *p = c;
-                            }
-                        }
-                        mask => {
-                            let channel = match mask {
-                                ImageMaskMode::A => 3,
-                                ImageMaskMode::B => 2,
-                                ImageMaskMode::G => 1,
-                                ImageMaskMode::R => 0,
-                                _ => unreachable!(),
-                            };
-                            for (p, rgba) in a.iter_mut().zip(raw.chunks_exact(4)) {
-                                let c = rgba[channel];
-                                if c < 255 {
-                                    is_opaque = false;
-                                }
-                                *p = c;
-                            }
-                        }
-                    }
-                    a
-                } else {
-                    let mut buf = raw;
-                    buf.chunks_mut(4).for_each(|c| {
-                        if c[3] < 255 {
-                            is_opaque = false;
-                            let a = c[3] as f32 / 255.0;
-                            c[0..3].iter_mut().for_each(|c| *c = (*c as f32 * a) as u8);
-                        }
-                        c.swap(0, 2);
-                    });
-                    IpcBytesMut::from_vec_blocking(buf)?
-                }
-            }
-            ImageLuma16(img) => {
-                let raw = img.into_raw();
-                if mask.is_some() {
-                    let mut a = IpcBytes::new_mut_blocking(pixels_len)?;
-                    for (p, l) in a.iter_mut().zip(raw) {
-                        let l = (l as f32 / u16::MAX as f32 * 255.0) as u8;
-                        if l < 255 {
-                            is_opaque = false;
-                        }
-                        *p = l;
-                    }
-                    a
-                } else {
-                    let mut bgra = IpcBytes::new_mut_blocking(pixels_len * 4)?;
-                    for (p, l) in bgra.chunks_exact_mut(4).zip(raw) {
-                        let l = (l as f32 / u16::MAX as f32 * 255.0) as u8;
-                        p.copy_from_slice(&[l, l, l, 255]);
-                    }
-                    bgra
-                }
-            }
-            ImageLumaA16(img) => {
-                let raw = img.into_raw();
-                if let Some(mask) = mask {
-                    let mut a = IpcBytes::new_mut_blocking(pixels_len)?;
-                    match mask {
-                        ImageMaskMode::A => {
-                            for (p, la) in a.iter_mut().zip(raw.chunks_exact(2)) {
-                                if la[1] < u16::MAX {
-                                    is_opaque = false;
-                                }
-                                let max = u16::MAX as f32;
-                                let l = la[1] as f32 / max * 255.0;
-                                *p = l as u8;
-                            }
-                        }
-                        ImageMaskMode::B | ImageMaskMode::G | ImageMaskMode::R | ImageMaskMode::Luminance => {
-                            for (p, la) in a.iter_mut().zip(raw.chunks_exact(2)) {
-                                if la[0] < u16::MAX {
-                                    is_opaque = false;
-                                }
-                                let max = u16::MAX as f32;
-                                let l = la[0] as f32 / max * 255.0;
-                                *p = l as u8;
-                            }
-                        }
-                        _ => unimplemented!(),
-                    }
-                    a
-                } else {
-                    let mut bgra = IpcBytes::new_mut_blocking(pixels_len * 4)?;
-                    for (p, la) in bgra.chunks_exact_mut(4).zip(raw.chunks_exact(2)) {
-                        let max = u16::MAX as f32;
-                        let l = la[0] as f32 / max;
-                        let a = la[1] as f32 / max * 255.0;
-
-                        p.copy_from_slice(&if la[1] < u16::MAX {
-                            is_opaque = false;
-                            let l = (l * a) as u8;
-                            [l, l, l, a as u8]
-                        } else {
-                            let l = (l * 255.0) as u8;
-                            [l, l, l, a as u8]
-                        });
-                    }
-                    bgra
-                }
-            }
-            ImageRgb16(img) => {
-                let raw = img.into_raw();
-                if let Some(mask) = mask {
-                    let mut a = IpcBytes::new_mut_blocking(pixels_len)?;
-                    match mask {
-                        ImageMaskMode::Luminance | ImageMaskMode::A => {
-                            for (p, rgb) in a.iter_mut().zip(raw.chunks_exact(3)) {
-                                let c = luminance_16(rgb);
-                                if c < 255 {
-                                    is_opaque = false;
-                                }
-                                *p = c;
-                            }
-                        }
-                        mask => {
-                            let channel = match mask {
-                                ImageMaskMode::B => 2,
-                                ImageMaskMode::G => 1,
-                                ImageMaskMode::R => 0,
-                                _ => unreachable!(),
-                            };
-                            for (p, rgb) in a.iter_mut().zip(raw.chunks_exact(3)) {
-                                let c = rgb[channel];
-                                if c < u16::MAX {
-                                    is_opaque = false;
-                                }
-                                *p = (c as f32 / u16::MAX as f32 * 255.0) as u8;
-                            }
-                        }
-                    }
-                    a
-                } else {
-                    let mut bgra = IpcBytes::new_mut_blocking(pixels_len * 4)?;
-                    for (p, rgb) in bgra.chunks_exact_mut(4).zip(raw.chunks_exact(3)) {
-                        let to_u8 = 255.0 / u16::MAX as f32;
-                        p.copy_from_slice(&[
-                            (rgb[2] as f32 * to_u8) as u8,
-                            (rgb[1] as f32 * to_u8) as u8,
-                            (rgb[0] as f32 * to_u8) as u8,
-                            255,
-                        ]);
-                    }
-                    bgra
-                }
-            }
-            ImageRgba16(img) => {
-                let raw = img.into_raw();
-                if let Some(mask) = mask {
-                    let mut a = IpcBytes::new_mut_blocking(pixels_len)?;
-                    match mask {
-                        ImageMaskMode::Luminance => {
-                            for (p, rgba) in a.iter_mut().zip(raw.chunks_exact(4)) {
-                                let c = luminance_16(&rgba[..3]);
-                                if c < 255 {
-                                    is_opaque = false;
-                                }
-                                *p = c;
-                            }
-                        }
-                        mask => {
-                            let channel = match mask {
-                                ImageMaskMode::A => 3,
-                                ImageMaskMode::B => 2,
-                                ImageMaskMode::G => 1,
-                                ImageMaskMode::R => 0,
-                                _ => unreachable!(),
-                            };
-                            for (p, rgba) in a.iter_mut().zip(raw.chunks_exact(4)) {
-                                let c = rgba[channel];
-                                if c < 255 {
-                                    is_opaque = false;
-                                }
-                                *p = (c as f32 / u16::MAX as f32 * 255.0) as u8;
-                            }
-                        }
-                    }
-                    a
-                } else {
-                    let mut bgra = IpcBytes::new_mut_blocking(pixels_len * 4)?;
-                    for (p, rgba) in bgra.chunks_exact_mut(4).zip(raw.chunks_exact(4)) {
-                        let c = if rgba[3] < u16::MAX {
-                            is_opaque = false;
-                            let max = u16::MAX as f32;
-                            let a = rgba[3] as f32 / max * 255.0;
-                            [
-                                (rgba[2] as f32 / max * a) as u8,
-                                (rgba[1] as f32 / max * a) as u8,
-                                (rgba[0] as f32 / max * a) as u8,
-                                a as u8,
-                            ]
-                        } else {
-                            let to_u8 = 255.0 / u16::MAX as f32;
-                            [
-                                (rgba[2] as f32 * to_u8) as u8,
-                                (rgba[1] as f32 * to_u8) as u8,
-                                (rgba[0] as f32 * to_u8) as u8,
-                                255,
-                            ]
-                        };
-                        p.copy_from_slice(&c);
-                    }
-                    bgra
-                }
-            }
-            ImageRgb32F(img) => {
-                let raw = img.into_raw();
-                if let Some(mask) = mask {
-                    let mut a = IpcBytes::new_mut_blocking(pixels_len)?;
-                    match mask {
-                        ImageMaskMode::Luminance | ImageMaskMode::A => {
-                            for (p, rgb) in a.iter_mut().zip(raw.chunks_exact(3)) {
-                                let c = luminance_f32(rgb);
-                                if c < 255 {
-                                    is_opaque = false;
-                                }
-                                *p = c;
-                            }
-                        }
-                        mask => {
-                            let channel = match mask {
-                                ImageMaskMode::B => 2,
-                                ImageMaskMode::G => 1,
-                                ImageMaskMode::R => 0,
-                                _ => unreachable!(),
-                            };
-                            for (p, rgb) in a.iter_mut().zip(raw.chunks_exact(3)) {
-                                let c = (rgb[channel] * 255.0) as u8;
-                                if c < 255 {
-                                    is_opaque = false;
-                                }
-                                *p = c;
-                            }
-                        }
-                    }
-                    a
-                } else {
-                    let mut bgra = IpcBytes::new_mut_blocking(pixels_len * 4)?;
-                    for (p, rgb) in bgra.chunks_exact_mut(4).zip(raw.chunks_exact(3)) {
-                        p.copy_from_slice(&[(rgb[2] * 255.0) as u8, (rgb[1] * 255.0) as u8, (rgb[0] * 255.0) as u8, 255]);
-                    }
-                    bgra
-                }
-            }
-            ImageRgba32F(img) => {
-                let raw = img.into_raw();
-                if let Some(mask) = mask {
-                    let mut a = IpcBytes::new_mut_blocking(pixels_len)?;
-                    match mask {
-                        ImageMaskMode::Luminance => {
-                            for (p, rgba) in a.iter_mut().zip(raw.chunks_exact(4)) {
-                                let c = luminance_f32(&rgba[..3]);
-                                if c < 255 {
-                                    is_opaque = false;
-                                }
-                                *p = c;
-                            }
-                        }
-                        mask => {
-                            let channel = match mask {
-                                ImageMaskMode::A => 3,
-                                ImageMaskMode::B => 2,
-                                ImageMaskMode::G => 1,
-                                ImageMaskMode::R => 0,
-                                _ => unreachable!(),
-                            };
-                            for (p, rgba) in a.iter_mut().zip(raw.chunks_exact(4)) {
-                                let c = (rgba[channel] * 255.0) as u8;
-                                if c < 255 {
-                                    is_opaque = false;
-                                }
-                                *p = c;
-                            }
-                        }
-                    }
-                    a
-                } else {
-                    let mut bgra = IpcBytes::new_mut_blocking(pixels_len * 4)?;
-                    for (p, rgba) in bgra.chunks_exact_mut(4).zip(raw.chunks_exact(4)) {
-                        let c = if rgba[3] < 1.0 {
-                            is_opaque = false;
-                            let a = rgba[3] * 255.0;
-                            [(rgba[2] * a) as u8, (rgba[1] * a) as u8, (rgba[0] * a) as u8, a as u8]
-                        } else {
-                            [(rgba[2] * 255.0) as u8, (rgba[1] * 255.0) as u8, (rgba[0] * 255.0) as u8, 255]
-                        };
-                        p.copy_from_slice(&c);
-                    }
-                    bgra
-                }
-            }
-            _ => unreachable!(),
-        };
-
-        #[cfg(feature = "image_any")]
-        if let Some(p) = icc_profile {
-            use lcms2::*;
-            let srgb = Profile::new_srgb();
-            let t = Transform::new(&p, PixelFormat::BGRA_8, &srgb, PixelFormat::BGRA_8, Intent::Perceptual).unwrap();
-            t.transform_in_place(&mut pixels);
-        }
-        #[cfg(not(feature = "image_any"))]
-        let _ = (icc_profile, &mut pixels);
-
-        Ok((
-            pixels.finish_blocking()?,
-            PxSize::new(Px(size.0 as i32), Px(size.1 as i32)),
-            density,
-            is_opaque,
-            mask.is_some(), // is_mask
-        ))
-    }
-
-    pub fn encode(&self, id: ImageId, format: Txt) {
-        if !ENCODERS.contains(&format.as_str()) {
-            let error = formatx!("cannot encode `{id:?}` to `{format}`, unknown format");
-            let _ = self
-                .app_sender
-                .send(AppEvent::Notify(Event::ImageEncodeError { image: id, format, error }));
-            return;
-        }
-
-        if let Some(img) = self.get(id) {
-            let fmt = image::ImageFormat::from_extension(format.as_str()).unwrap();
-            debug_assert!(fmt.can_write());
-
-            let img = img.clone();
-            let sender = self.app_sender.clone();
-            rayon::spawn(move || {
-                let mut data = IpcBytes::new_writer_blocking();
-                match img.encode(fmt, &mut data) {
-                    Ok(_) => match data.finish() {
-                        Ok(data) => {
-                            let _ = sender.send(AppEvent::Notify(Event::ImageEncoded { image: id, format, data }));
-                        }
-                        Err(e) => {
-                            let _ = sender.send(AppEvent::Notify(Event::ImageEncodeError {
-                                image: id,
-                                format,
-                                error: e.to_txt(),
-                            }));
-                        }
-                    },
-                    Err(e) => {
-                        let error = formatx!("failed to encode `{id:?}` to `{format}`, {e}");
-                        let _ = sender.send(AppEvent::Notify(Event::ImageEncodeError { image: id, format, error }));
-                    }
-                }
-            })
-        } else {
-            let error = formatx!("cannot encode `{id:?}` to `{format}`, image not found");
-            let _ = self
-                .app_sender
-                .send(AppEvent::Notify(Event::ImageEncodeError { image: id, format, error }));
-        }
-    }
-
     pub(crate) fn on_low_memory(&mut self) {
         // app-process controls this cache
     }
@@ -1203,176 +497,6 @@ impl Image {
         }
     }
 
-    /// Generate a window icon from the image.
-    pub fn icon(&self) -> Option<Icon> {
-        let (size, pixels) = match &*self.0 {
-            ImageData::RawData { size, pixels, .. } => (size, pixels),
-            ImageData::NativeTexture { .. } => unreachable!(),
-        };
-
-        let width = size.width.0 as u32;
-        let height = size.height.0 as u32;
-        if width == 0 || height == 0 || self.0.is_mask() {
-            None
-        } else {
-            let r = if width > 255 || height > 255 {
-                // resize to max 255
-                let mut buf = pixels[..].to_vec();
-                // BGRA to RGBA
-                buf.chunks_exact_mut(4).for_each(|c| c.swap(0, 2));
-                let img = image::ImageBuffer::from_raw(width, height, buf).unwrap();
-                let img = image::DynamicImage::ImageRgba8(img);
-                let img = img.resize(255, 255, image::imageops::FilterType::Lanczos3);
-
-                use image::GenericImageView;
-                let (width, height) = img.dimensions();
-                let buf = img.into_rgba8().into_raw();
-                winit::window::Icon::from_rgba(buf, width, height)
-            } else {
-                let mut buf = pixels[..].to_vec();
-                // BGRA to RGBA
-                buf.chunks_exact_mut(4).for_each(|c| c.swap(0, 2));
-                winit::window::Icon::from_rgba(buf, width, height)
-            };
-            match r {
-                Ok(i) => Some(i),
-                Err(e) => {
-                    tracing::error!("failed to convert image to custom icon, {e}");
-                    None
-                }
-            }
-        }
-    }
-
-    /// Generate a cursor from the image.
-    pub fn cursor(&self, hotspot: PxPoint, event_loop: &ActiveEventLoop) -> Option<CustomCursor> {
-        let (size, pixels) = match &*self.0 {
-            ImageData::RawData { size, pixels, .. } => (size, pixels),
-            ImageData::NativeTexture { .. } => unreachable!(),
-        };
-
-        let width: u16 = size.width.0.try_into().ok()?;
-        let height: u16 = size.height.0.try_into().ok()?;
-        let hotspot_x: u16 = hotspot.x.0.try_into().ok()?;
-        let hotspot_y: u16 = hotspot.y.0.try_into().ok()?;
-
-        if width == 0 || height == 0 || hotspot_x > width || hotspot_y > height || self.0.is_mask() {
-            None
-        } else {
-            let mut buf = pixels[..].to_vec();
-            // BGRA to RGBA
-            buf.chunks_exact_mut(4).for_each(|c| c.swap(0, 2));
-            match CustomCursor::from_rgba(buf, width, height, hotspot_x, hotspot_y) {
-                Ok(c) => Some(event_loop.create_custom_cursor(c)),
-                Err(e) => {
-                    tracing::error!("failed to convert image to custom cursor, {e}");
-                    None
-                }
-            }
-        }
-    }
-
-    pub fn encode(&self, format: image::ImageFormat, buffer: &mut (impl std::io::Write + std::io::Seek)) -> image::ImageResult<()> {
-        let (size, pixels, density) = match &*self.0 {
-            ImageData::RawData { size, pixels, density, .. } => (size, pixels, density),
-            ImageData::NativeTexture { .. } => unreachable!(),
-        };
-
-        if size.width <= 0 || size.height <= 0 {
-            return Err(image::ImageError::IoError(std::io::Error::new(
-                std::io::ErrorKind::InvalidInput,
-                "cannot encode zero sized image",
-            )));
-        }
-
-        if self.0.is_mask() {
-            let width = size.width.0 as u32;
-            let height = size.height.0 as u32;
-            let is_opaque = self.0.is_opaque();
-            let r8 = pixels[..].to_vec();
-
-            let mut img = image::DynamicImage::ImageLuma8(image::ImageBuffer::from_raw(width, height, r8).unwrap());
-            if is_opaque {
-                img = image::DynamicImage::ImageRgb8(img.to_rgb8());
-            }
-            img.write_to(buffer, format)?;
-
-            return Ok(());
-        }
-
-        // invert rows, `image` only supports top-to-bottom buffers.
-        let mut buf = pixels[..].to_vec();
-        // BGRA to RGBA
-        buf.chunks_exact_mut(4).for_each(|c| c.swap(0, 2));
-        let rgba = buf;
-
-        let width = size.width.0 as u32;
-        let height = size.height.0 as u32;
-        let is_opaque = self.0.is_opaque();
-
-        match format {
-            #[cfg(feature = "image_jpeg")]
-            image::ImageFormat::Jpeg => {
-                let mut jpg = image::codecs::jpeg::JpegEncoder::new(buffer);
-                if let Some(density) = density {
-                    jpg.set_pixel_density(image::codecs::jpeg::PixelDensity {
-                        density: (density.height.ppi() as u16, density.height.ppi() as u16),
-                        unit: image::codecs::jpeg::PixelDensityUnit::Inches,
-                    });
-                }
-                jpg.encode(&rgba, width, height, image::ColorType::Rgba8.into())?;
-            }
-            #[cfg(feature = "image_png")]
-            image::ImageFormat::Png => {
-                let mut img = image::DynamicImage::ImageRgba8(image::ImageBuffer::from_raw(width, height, rgba).unwrap());
-                if is_opaque {
-                    img = image::DynamicImage::ImageRgb8(img.to_rgb8());
-                }
-                if let Some(density) = density {
-                    let mut png_bytes = vec![];
-
-                    img.write_to(&mut std::io::Cursor::new(&mut png_bytes), image::ImageFormat::Png)?;
-
-                    let mut png = img_parts::png::Png::from_bytes(png_bytes.into()).unwrap();
-
-                    let chunk_kind = *b"pHYs";
-                    debug_assert!(png.chunk_by_type(chunk_kind).is_none());
-
-                    use byteorder::*;
-                    let mut chunk = Vec::with_capacity(4 * 2 + 1);
-
-                    // density / inch_to_metric
-                    let ppm_x = density.width.ppm() as u32;
-                    let ppm_y = density.height.ppm() as u32;
-
-                    chunk.write_u32::<BigEndian>(ppm_x).unwrap();
-                    chunk.write_u32::<BigEndian>(ppm_y).unwrap();
-                    chunk.write_u8(1).unwrap(); // metric
-
-                    let chunk = img_parts::png::PngChunk::new(chunk_kind, chunk.into());
-                    png.chunks_mut().insert(1, chunk);
-
-                    png.encoder().write_to(buffer)?;
-                } else {
-                    img.write_to(buffer, image::ImageFormat::Png)?;
-                }
-            }
-            _ => {
-                // other formats that we don't with custom PPI meta.
-
-                let _ = density; // suppress warning when both png an jpeg are not enabled
-
-                let mut img = image::DynamicImage::ImageRgba8(image::ImageBuffer::from_raw(width, height, rgba).unwrap());
-                if is_opaque {
-                    img = image::DynamicImage::ImageRgb8(img.to_rgb8());
-                }
-                img.write_to(buffer, format)?;
-            }
-        }
-
-        Ok(())
-    }
-
     #[allow(unused)]
     pub fn size(&self) -> PxSize {
         match &*self.0 {
@@ -1388,350 +512,4 @@ impl Image {
             ImageData::NativeTexture { .. } => unreachable!(),
         }
     }
-}
-
-// Image data is provided to webrender directly from the BGRA8 shared memory.
-// The `ExternalImageId` is the Arc pointer to ImageData.
-mod external {
-    use std::{collections::hash_map::Entry, sync::Arc};
-
-    use rustc_hash::FxHashMap;
-    use webrender::{
-        RenderApi,
-        api::{
-            DocumentId, ExternalImage, ExternalImageData, ExternalImageHandler, ExternalImageId, ExternalImageSource, ExternalImageType,
-            ImageKey,
-            units::{ImageDirtyRect, TexelRect},
-        },
-    };
-    use zng_view_api::image::ImageTextureId;
-
-    use super::{Image, ImageData};
-
-    /// Implements [`ExternalImageHandler`].
-    ///
-    /// # Safety
-    ///
-    /// This is only safe if use with [`ImageUseMap`].
-    pub(crate) struct WrImageCache {
-        locked: Vec<Arc<ImageData>>,
-    }
-    impl WrImageCache {
-        pub fn new_boxed() -> Box<dyn ExternalImageHandler> {
-            Box::new(WrImageCache { locked: vec![] })
-        }
-    }
-    impl ExternalImageHandler for WrImageCache {
-        fn lock(&mut self, key: ExternalImageId, _channel_index: u8, _is_composited: bool) -> ExternalImage<'_> {
-            // SAFETY: this is safe because the Arc is kept alive in `ImageUseMap`.
-            let img = unsafe {
-                let ptr = key.0 as *const ImageData;
-                Arc::increment_strong_count(ptr);
-                Arc::<ImageData>::from_raw(ptr)
-            };
-
-            self.locked.push(img); // keep alive in case the image is removed mid-use
-
-            match &**self.locked.last().unwrap() {
-                ImageData::RawData { pixels, .. } => {
-                    ExternalImage {
-                        uv: TexelRect::invalid(), // `RawData` does not use `uv`.
-                        source: ExternalImageSource::RawData(&pixels[..]),
-                    }
-                }
-                ImageData::NativeTexture { uv, texture: id } => ExternalImage {
-                    uv: *uv,
-                    source: ExternalImageSource::NativeTexture(*id),
-                },
-            }
-        }
-
-        fn unlock(&mut self, key: ExternalImageId, _channel_index: u8) {
-            if let Some(i) = self.locked.iter().position(|d| ExternalImageId(Arc::as_ptr(d) as _) == key) {
-                self.locked.swap_remove(i);
-            } else {
-                debug_assert!(false);
-            }
-        }
-    }
-
-    impl Image {
-        fn external_id(&self) -> ExternalImageId {
-            ExternalImageId(Arc::as_ptr(&self.0) as u64)
-        }
-
-        fn data(&self) -> webrender::api::ImageData {
-            webrender::api::ImageData::External(ExternalImageData {
-                id: self.external_id(),
-                channel_index: 0,
-                image_type: ExternalImageType::Buffer,
-                normalized_uvs: false,
-            })
-        }
-    }
-
-    /// Track and manage images used in a renderer.
-    ///
-    /// The renderer must use [`WrImageCache`] as the external image source.
-    #[derive(Default)]
-    pub(crate) struct ImageUseMap {
-        id_tex: FxHashMap<ExternalImageId, (ImageTextureId, Image)>,
-        tex_id: FxHashMap<ImageTextureId, ExternalImageId>,
-    }
-    impl ImageUseMap {
-        pub fn new_use(&mut self, image: &Image, document_id: DocumentId, api: &mut RenderApi) -> ImageTextureId {
-            let id = image.external_id();
-            match self.id_tex.entry(id) {
-                Entry::Occupied(e) => e.get().0,
-                Entry::Vacant(e) => {
-                    let key = api.generate_image_key();
-                    let tex_id = ImageTextureId::from_raw(key.1);
-                    e.insert((tex_id, image.clone())); // keep the image Arc alive, we expect this in `WrImageCache`.
-                    self.tex_id.insert(tex_id, id);
-
-                    let mut txn = webrender::Transaction::new();
-                    txn.add_image(key, image.descriptor(), image.data(), None);
-                    api.send_transaction(document_id, txn);
-
-                    tex_id
-                }
-            }
-        }
-
-        /// Returns if needs to update.
-        pub fn update_use(&mut self, texture_id: ImageTextureId, image: &Image, document_id: DocumentId, api: &mut RenderApi) {
-            if let Entry::Occupied(mut e) = self.tex_id.entry(texture_id) {
-                let id = image.external_id();
-                if *e.get() != id {
-                    let prev_id = e.insert(id);
-                    self.id_tex.remove(&prev_id).unwrap();
-                    self.id_tex.insert(id, (texture_id, image.clone()));
-
-                    let mut txn = webrender::Transaction::new();
-                    txn.update_image(
-                        ImageKey(api.get_namespace_id(), texture_id.get()),
-                        image.descriptor(),
-                        image.data(),
-                        &ImageDirtyRect::All,
-                    );
-                    api.send_transaction(document_id, txn);
-                }
-            }
-        }
-
-        pub fn delete(&mut self, texture_id: ImageTextureId, document_id: DocumentId, api: &mut RenderApi) {
-            if let Some(id) = self.tex_id.remove(&texture_id) {
-                let _img = self.id_tex.remove(&id); // remove but keep alive until the transaction is done.
-                let mut txn = webrender::Transaction::new();
-                txn.delete_image(ImageKey(api.get_namespace_id(), texture_id.get()));
-                api.send_transaction(document_id, txn);
-            }
-        }
-    }
-}
-pub(crate) use external::{ImageUseMap, WrImageCache};
-
-mod capture {
-    use std::sync::Arc;
-
-    use webrender::api::{ImageDescriptor, ImageDescriptorFlags, ImageFormat};
-    use zng_task::channel::IpcBytes;
-    use zng_txt::formatx;
-    use zng_unit::{Factor, PxDensity2d, PxDensityUnits as _, PxRect};
-    use zng_view_api::{
-        Event,
-        image::{ImageDataFormat, ImageId, ImageLoadedData, ImageMaskMode, ImageRequest},
-        window::{FrameId, WindowId},
-    };
-
-    use crate::{
-        AppEvent,
-        image_cache::{Image, ImageData},
-    };
-
-    use super::ImageCache;
-
-    impl ImageCache {
-        /// Create frame_image for an `Api::frame_image` request.
-        pub fn frame_image(
-            &mut self,
-            gl: &dyn gleam::gl::Gl,
-            rect: PxRect,
-            window_id: WindowId,
-            frame_id: FrameId,
-            scale_factor: Factor,
-            mask: Option<ImageMaskMode>,
-        ) -> ImageId {
-            if frame_id == FrameId::INVALID {
-                let id = self.image_id_gen.incr();
-                let _ = self.app_sender.send(AppEvent::Notify(Event::ImageLoadError {
-                    image: id,
-                    error: formatx!("no frame rendered in window `{window_id:?}`"),
-                }));
-                let _ = self.app_sender.send(AppEvent::Notify(Event::FrameImageReady {
-                    window: window_id,
-                    frame: frame_id,
-                    image: id,
-                    selection: rect,
-                }));
-                return id;
-            }
-
-            let data = self.frame_image_data(gl, rect, scale_factor, mask);
-
-            let id = data.id;
-
-            let _ = self.app_sender.send(AppEvent::ImageLoaded(data));
-            let _ = self.app_sender.send(AppEvent::Notify(Event::FrameImageReady {
-                window: window_id,
-                frame: frame_id,
-                image: id,
-                selection: rect,
-            }));
-
-            id
-        }
-
-        /// Create frame_image for a capture request in the FrameRequest.
-        pub fn frame_image_data(
-            &mut self,
-            gl: &dyn gleam::gl::Gl,
-            rect: PxRect,
-            scale_factor: Factor,
-            mask: Option<ImageMaskMode>,
-        ) -> ImageLoadedData {
-            let data = self.frame_image_data_impl(gl, rect, scale_factor, mask);
-
-            let flags = if data.is_opaque {
-                ImageDescriptorFlags::IS_OPAQUE
-            } else {
-                ImageDescriptorFlags::empty()
-            };
-
-            self.images.insert(
-                data.id,
-                Image(Arc::new(ImageData::RawData {
-                    size: data.size,
-                    pixels: data.pixels.clone(),
-                    descriptor: ImageDescriptor::new(
-                        data.size.width.0,
-                        data.size.height.0,
-                        if data.is_mask { ImageFormat::R8 } else { ImageFormat::BGRA8 },
-                        flags,
-                    ),
-                    density: data.density,
-                })),
-            );
-
-            data
-        }
-
-        fn frame_image_data_impl(
-            &mut self,
-            gl: &dyn gleam::gl::Gl,
-            rect: PxRect,
-            scale_factor: Factor,
-            mask: Option<ImageMaskMode>,
-        ) -> ImageLoadedData {
-            let format = match gl.get_type() {
-                gleam::gl::GlType::Gl => gleam::gl::BGRA,
-                gleam::gl::GlType::Gles => gleam::gl::RGBA,
-            };
-            let pixels_flipped = gl.read_pixels(
-                rect.origin.x.0,
-                rect.origin.y.0,
-                rect.size.width.0,
-                rect.size.height.0,
-                format,
-                gleam::gl::UNSIGNED_BYTE,
-            );
-            let mut buf = vec![0u8; pixels_flipped.len()];
-            assert_eq!(rect.size.width.0 as usize * rect.size.height.0 as usize * 4, buf.len());
-            let stride = 4 * rect.size.width.0 as usize;
-            for (px, buf) in pixels_flipped.chunks_exact(stride).rev().zip(buf.chunks_exact_mut(stride)) {
-                buf.copy_from_slice(px);
-            }
-
-            if let Some(mask) = mask {
-                if format == gleam::gl::BGRA {
-                    for bgra in buf.chunks_exact_mut(4) {
-                        bgra.swap(0, 3);
-                    }
-                }
-                let density = 96.0 * scale_factor.0;
-                let density = Some(PxDensity2d::splat(density.ppi()));
-
-                let (pixels, size, density, is_opaque, is_mask) = Self::convert_decoded(
-                    image::DynamicImage::ImageRgba8(
-                        image::ImageBuffer::from_raw(rect.size.width.0 as u32, rect.size.height.0 as u32, buf).unwrap(),
-                    ),
-                    Some(mask),
-                    density,
-                    None,
-                )
-                .unwrap(); // frame size is not large enough to trigger an memmap that can fail
-
-                let id = self.add(ImageRequest::new(
-                    ImageDataFormat::A8 { size },
-                    pixels.clone(),
-                    u64::MAX,
-                    None,
-                    Some(mask),
-                ));
-
-                ImageLoadedData::new(id, size, density, is_opaque, is_mask, pixels)
-            } else {
-                if format == gleam::gl::RGBA {
-                    for rgba in buf.chunks_exact_mut(4) {
-                        rgba.swap(0, 3);
-                    }
-                }
-
-                let is_opaque = buf.chunks_exact(4).all(|bgra| bgra[3] == 255);
-
-                let data = IpcBytes::from_vec_blocking(buf).unwrap(); // frame size is not large enough to trigger an memmap that can fail
-                let density = 96.0 * scale_factor.0;
-                let density = Some(PxDensity2d::splat(density.ppi()));
-                let size = rect.size;
-
-                let id = self.add(ImageRequest::new(
-                    ImageDataFormat::Bgra8 { size, density },
-                    data.clone(),
-                    u64::MAX,
-                    None,
-                    mask,
-                ));
-
-                ImageLoadedData::new(id, size, density, is_opaque, false, data)
-            }
-        }
-    }
-}
-
-fn luminance(rgb: &[u8]) -> u8 {
-    let r = rgb[0] as f32 / 255.0;
-    let g = rgb[1] as f32 / 255.0;
-    let b = rgb[2] as f32 / 255.0;
-
-    let l = r * 0.2126 + g * 0.7152 + b * 0.0722;
-    (l * 255.0) as u8
-}
-
-fn luminance_16(rgb: &[u16]) -> u8 {
-    let max = u16::MAX as f32;
-    let r = rgb[0] as f32 / max;
-    let g = rgb[1] as f32 / max;
-    let b = rgb[2] as f32 / max;
-
-    let l = r * 0.2126 + g * 0.7152 + b * 0.0722;
-    (l * 255.0) as u8
-}
-
-fn luminance_f32(rgb: &[f32]) -> u8 {
-    let r = rgb[0];
-    let g = rgb[1];
-    let b = rgb[2];
-
-    let l = r * 0.2126 + g * 0.7152 + b * 0.0722;
-    (l * 255.0) as u8
 }
